% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getexchrate_av.R
\name{getexchrate_av}
\alias{getexchrate_av}
\title{getexchrate_av}
\usage{
getexchrate_av(to_curr, from_curr = "USD")
}
\arguments{
\item{from_curr}{}
}
\value{
data.frame
}
\description{
API wrapper to extract exchange rate data from www.alphavantage.co
}
