% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getstockprice_ms.R
\name{getstockprice_ms}
\alias{getstockprice_ms}
\title{getstockprice_ms}
\usage{
getstockprice_ms(symbol)
}
\arguments{
\item{symbol}{stock symbol in the stock market}
}
\value{

}
\description{
getstockprice_ms
}
